backend/requirements.txt
discord.py === 2.5.2
Django==5.1.7
djangorestframework==3.16.0
gunicorn === 23.0.0
requests === 2.32.3
psycopg2-binary === 2.9.10